Add options for blenderrelease paths. Print the path that is used to gather all addons. Use persistent instead of temporary path.

+++ gather_and_link_addons.sh
@@ -6,7 +6,10 @@ if [ ! -z $PATH_TO_BLENDER_RELEASE ]; then
 elif [ -z $BLENDER_RELEASE_PATH ]; then
 	echo 'Using user provided blender release: '$BLENDER_RELEASE_PATH
 else
+#    BLENDER_RELEASE_PATH=~/blender
+#    BLENDER_RELEASE_PATH=~/blenderart
     BLENDER_RELEASE_PATH=~/blendercad
+#    BLENDER_RELEASE_PATH=~/blendercam
 	echo 'Using default blender release location: '$BLENDER_RELEASE_PATH
 fi
 
@@ -15,13 +18,15 @@ BLENDER_ADDONS_PATH=$BLENDER_SCRIPTS_PATH/addons
 BLENDER_ADDONS_BACKUP_PATH=$BLENDER_ADDONS_PATH'.bak'
 
 CONFIG_PATH=~/.config/blender
-PATH_FOR_TEMPORARY_FILES=$CONFIG_PATH
-if [ ! -d $CONFIG_PATH ]; then
-	PATH_FOR_TEMPORARY_FILES=/tmp
-fi
+#PATH_FOR_TEMPORARY_FILES=/tmp
 
 # For collecting all addons from all the repos into one folder:
-ADDONS_COLLECTION_PATH=$PATH_FOR_TEMPORARY_FILES/addons__only_put_here_links
+ADDONS_COLLECTION_FOLDERNAME=addons__only_put_here_links
+if [ ! -d $CONFIG_PATH ]; then
+	ADDONS_COLLECTION_PATH=$HOME/$ADDONS_COLLECTION_FOLDERNAME
+else
+	ADDONS_COLLECTION_PATH=$CONFIG_PATH/$ADDONS_COLLECTION_FOLDERNAME
+fi
 
 # Remove an out to date link:
 # If it's a directory, backup: (and copy back later such that the custom addons take precedence/override the default ones and yet special addons that are only found in a release like cycles/ exist.)
@@ -33,6 +38,7 @@ elif ( -f $BLENDER_ADDONS_PATH ); then
     rm $BLENDER_ADDONS_PATH
 fi
 
+echo 'Linking addons collection path into '
 ln -s $ADDONS_COLLECTION_PATH $BLENDER_ADDONS_PATH
 
 if (( -f $ADDONS_COLLECTION_PATH )); then
@@ -44,12 +50,17 @@ elif ! (( -d $ADDONS_COLLECTION_PATH )); then
 fi
 
 # Remove all links as we will relink anew later:
-rm $ADDONS_COLLECTION_PATH/*
+# Prevent deleting all home directory files:
+if [ ! -z $ADDONS_COLLECTION_PATH ]; then
+    rm $ADDONS_COLLECTION_PATH/*
+fi
 
 # Note that space ' ' is delimiter by default:
 # Order matters, later mentioned folders can't override addons that already exist.
 if [ ! -z $ALL_PATHS_TO_ADDONS ]; then
 	echo 'Using user provided blender addons: '$ALL_PATHS_TO_ADDONS
+	echo 'Appending previously existing addons, some are required e.g. cycles:'
+	$ALL_PATHS_TO_ADDONS="$ALL_PATHS_TO_ADDONS "$BLENDER_ADDONS_BACKUP_PATH
 else
 ALL_PATHS_TO_ADDONS=(
 
@@ -87,4 +98,5 @@ for i in "${ALL_PATHS_TO_ADDONS[@]}"; do
 	
 done
 
+echo '-> into ' ${ADDONS_COLLECTION_PATH}
